chore: restructure utils code
