+++ web3.js/src/index.ts
@@ -4,6 +4,7 @@ export * from './bpf-loader-deprecated';
 export * from './bpf-loader';
 export * from './connection';
 export * from './epoch-schedule';
+export * from './errors';
 export * from './fee-calculator';
 export * from './keypair';
 export * from './loader';
@@ -15,12 +16,7 @@ export * from './transaction';
 export * from './validator-info';
 export * from './vote-account';
 export * from './sysvar';
-export * from './errors';
-export * from './util/borsh-schema';
-export * from './util/send-and-confirm-transaction';
-export * from './util/send-and-confirm-raw-transaction';
-export * from './util/tx-expiry-custom-errors';
-export * from './util/cluster';
+export * from './utils';
 
 /**
  * There are 1-billion lamports in one SOL
