+++ web3.js/src/loader.ts
@@ -4,8 +4,8 @@ import * as BufferLayout from '@solana/buffer-layout';
 import {PublicKey} from './publickey';
 import {Transaction, PACKET_DATA_SIZE} from './transaction';
 import {SYSVAR_RENT_PUBKEY} from './sysvar';
-import {sendAndConfirmTransaction} from './util/send-and-confirm-transaction';
-import {sleep} from './util/sleep';
+import {sendAndConfirmTransaction} from './utils/send-and-confirm-transaction';
+import {sleep} from './utils/sleep';
 import type {Connection} from './connection';
 import type {Signer} from './keypair';
 import {SystemProgram} from './programs/system';
