+++ web3.js/src/programs/address-lookup-table.ts
@@ -3,7 +3,7 @@ import * as BufferLayout from '@solana/buffer-layout';
 
 import * as Layout from '../layout';
 import {PublicKey} from '../publickey';
-import * as bigintLayout from '../util/bigint';
+import * as bigintLayout from '../utils/bigint';
 import {SystemProgram} from './system';
 import {TransactionInstruction} from '../transaction';
 import {decodeData, encodeData, IInstructionInputData} from '../instruction';
