+++ web3.js/src/programs/stake.ts
@@ -15,7 +15,7 @@ import {
   SYSVAR_STAKE_HISTORY_PUBKEY,
 } from '../sysvar';
 import {Transaction, TransactionInstruction} from '../transaction';
-import {toBuffer} from '../util/to-buffer';
+import {toBuffer} from '../utils/to-buffer';
 
 /**
  * Address of the stake config account which configures the rate
