+++ web3.js/src/transaction/index.ts
@@ -1,2 +1,3 @@
 export * from './constants';
+export * from './expiry-custom-errors';
 export * from './legacy';
