+++ web3.js/src/transaction/legacy.ts
@@ -6,9 +6,9 @@ import {PACKET_DATA_SIZE, SIGNATURE_LENGTH_IN_BYTES} from './constants';
 import {Connection} from '../connection';
 import {Message} from '../message';
 import {PublicKey} from '../publickey';
-import * as shortvec from '../util/shortvec-encoding';
-import {toBuffer} from '../util/to-buffer';
-import invariant from '../util/assert';
+import * as shortvec from '../utils/shortvec-encoding';
+import {toBuffer} from '../utils/to-buffer';
+import invariant from '../utils/assert';
 import type {Signer} from '../keypair';
 import type {Blockhash} from '../blockhash';
 import type {CompiledInstruction} from '../message';
