+++ web3.js/src/utils/bigint.ts
