+++ web3.js/test/transaction-payer.test.ts
@@ -8,7 +8,7 @@ import {
   SystemProgram,
   LAMPORTS_PER_SOL,
 } from '../src';
-import invariant from '../src/util/assert';
+import invariant from '../src/utils/assert';
 import {MOCK_PORT, url} from './url';
 import {helpers, mockRpcResponse, mockServer} from './mocks/rpc-http';
 import {stubRpcWebSocket, restoreRpcWebSocket} from './mocks/rpc-websockets';
