+++ web3.js/test/transaction.test.ts
@@ -9,8 +9,8 @@ import {PublicKey} from '../src/publickey';
 import {Transaction, TransactionInstruction} from '../src/transaction';
 import {StakeProgram, SystemProgram} from '../src/programs';
 import {Message} from '../src/message';
-import invariant from '../src/util/assert';
-import {toBuffer} from '../src/util/to-buffer';
+import invariant from '../src/utils/assert';
+import {toBuffer} from '../src/utils/to-buffer';
 import {helpers} from './mocks/rpc-http';
 import {url} from './url';
